+++ tests/functional_tests/conftest.py
@@ -46,7 +46,7 @@ def no_use_catalog_cache(
     if "use_cache" in request.keywords:
         return
     monkeypatch.setattr("harlequin.app.get_catalog_cache", lambda *_: None)
-    monkeypatch.setattr("harlequin.app.update_catalog_cache", lambda *_: None)
+    monkeypatch.setattr("harlequin.app.update_catalog_cache", lambda *_, **__: None)
 
 
 @pytest.fixture(autouse=True)